# hll: skip recalc of sparseSet if tmpSet is empty